requirements.txt
Flask==2.0.1
RPi.GPIO==0.7.0
Flask-SQLAlchemy==2.5.1
python-dotenv==0.19.0